chore: reorganize utilities, add visuallyHidden util
