+++ src/shared/utilities/hasWindow.test.tsx
@@ -1,4 +1,4 @@
-import { hasWindow } from './utilities';
+import { hasWindow } from './hasWindow';
 
 describe('hasWindow', () => {
     test('should return true when the function is called in the browser', () => {
