+++ src/shared/utilities/hasWindow.ts
@@ -1,0 +1,7 @@
+/**
+ * Utility to verify if the window object exists
+ * @returns {boolean}
+ */
+export const hasWindow = (): boolean => {
+    return typeof window !== 'undefined';
+};
