+++ src/shared/utilities/visuallyHidden.ts
@@ -1,0 +1,13 @@
+const visuallyHidden: import('react').CSSProperties = {
+    border: 0,
+    clip: 'rect(0 0 0 0)',
+    height: '1px',
+    margin: -1,
+    overflow: 'hidden',
+    padding: 0,
+    position: 'absolute',
+    whiteSpace: 'nowrap',
+    width: '1px',
+};
+
+export default visuallyHidden;
